Ejercicios correctos

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: maialen <maialen@student.42.fr>            +#+  +:+       +#+        */
+/*   By: mmurgia- <mmurgia-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/19 11:07:31 by mmurgia-          #+#    #+#             */
-/*   Updated: 2022/05/31 13:18:16 by maialen          ###   ########.fr       */
+/*   Updated: 2022/06/07 11:43:44 by mmurgia-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@
 # include <string.h>
 # include <stdlib.h>
 # include <strings.h>
+# include <xlocale.h>
 # include <stdarg.h>
 # include <limits.h>
-# include <stdint.h>
 
 typedef struct s_list
 {
@@ -28,13 +28,13 @@ typedef struct s_list
 	struct s_list	*next;
 }					t_list;
 
-int		ft_printf(char const *str, ...);
-int		ft_format(va_list args, char type);
-int		ft_print_char(int c);
-int		ft_print_str(char *str);
-int		ft_print_num(int n);
-int		ft_print_percentage(void);
-int		ft_print_hex(unsigned long long int num, char c);
-int		ft_print_unsigned_num(unsigned int n);
+int	ft_printf(char const *str, ...);
+int	ft_format(va_list args, char type);
+int	ft_print_char(int c);
+int	ft_print_str(char *str);
+int	ft_print_num(int n);
+int	ft_print_percentage(void);
+int	ft_print_hex(unsigned long int num, char c);
+int	ft_print_unsigned_num(unsigned int n);
 
 #endif
